requirements.txt
# Core dependencies
click>=8.0.0
pydantic>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
requests>=2.30.0

# OpenAI integration
openai>=1.0.0

# Weaviate integration
weaviate-client>=4.0.0

# Job matching and web search
exa-py>=1.0.0

# Data handling
dataclasses-json>=0.6.0

# Utilities
rich>=13.0.0  # For better CLI output
tenacity>=8.0.0  # For retry logic

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0